+++ Firmware/drivers/Inc/common.h
@@ -3,6 +3,8 @@
 #include "stm32xx_hal.h"
 #include "stdint.h"
 
+void blinky_gpio_init(void);
+
 void gpio_clock_init(uint32_t port);
 
 void gpio_pin_init(uint32_t port, uint32_t pin);
